+++ src/chainparams.cpp
@@ -124,6 +124,9 @@ public:
         // Nov 15, 2019 12:00:00 UTC protocol upgrade
         consensus.gravitonActivationTime = 1573819200;
 
+        // May 15, 2020 12:00:00 UTC protocol upgrade
+        consensus.phononActivationTime = 1589544000;
+
         /**
          * The message start string is designed to be unlikely to occur in
          * normal data. The characters are rarely used upper ASCII, not valid as
@@ -287,6 +290,9 @@ public:
         // Nov 15, 2019 12:00:00 UTC protocol upgrade
         consensus.gravitonActivationTime = 1573819200;
 
+        // May 15, 2020 12:00:00 UTC protocol upgrade
+        consensus.phononActivationTime = 1589544000;
+
         diskMagic[0] = 0x0b;
         diskMagic[1] = 0x11;
         diskMagic[2] = 0x09;
@@ -401,6 +407,9 @@ public:
 
         // Nov 15, 2019 12:00:00 UTC protocol upgrade
         consensus.gravitonActivationTime = 1573819200;
+
+        // May 15, 2020 12:00:00 UTC protocol upgrade
+        consensus.phononActivationTime = 1589544000;
 
         diskMagic[0] = 0xfa;
         diskMagic[1] = 0xbf;