+++ doc/release-notes.md
@@ -6,3 +6,4 @@ This release includes the following features and fixes:
  - Support for Nov 2019 upgrade features, as detailed at https://github.com/bitcoincashorg/bitcoincash.org/blob/master/spec/2019-11-15-upgrade.md
     - Schnorr signatures for OP_CHECKMULTISIG(VERIFY): https://github.com/bitcoincashorg/bitcoincash.org/blob/master/spec/2019-11-15-schnorrmultisig.md
     - Enforce MINIMALDATA in script: https://github.com/bitcoincashorg/bitcoincash.org/blob/master/spec/2019-11-15-minimaldata.md
+ - Bump automatic replay protection to May 2020 upgrade.